From:     lane (Charles Lane) @ DUPHY4.Physics.Drexel.Edu Subject:  [patch URI 1.05] VMS mods To:       gisle @ aas.no Date:     Mon, 3 Apr 2000 10:58:46 EDT
The following is a patch to URI 1.05 to allow a clean test/install on
VMS machines.  There's only one minor change to the actual code: in
URI/file.pm the cwd is converted to "unix style" (Perl on VMS can use
either native-VMS or unix styles).

The remainder of the changes are to t/old-base.t, generally changing
native-VMS to unix styles after system() calls, and skipping tests
that involve having a chdir to /, which isn't valid on VMS.