From: Ed Avis <ed@membled.com> Subject: Re: URI::Heuristic: should it call hostfqdn() on loading? To: Gisle Aas <gisle@activestate.com> cc: libwww@perl.org Date: Fri, 6 Sep 2002 10:15:09 +0100 (BST)
This patch stops URI::Heuristic from accessing the network when it is
first loaded.  Instead, the way to access the two-letter country code
is through the MY_COUNTRY() function, which finds out the information
when it is called and caches it for later.  Setting
$URI::Heuristic::MY_COUNTRY by hand (as used in the test suite) still
works, but applications should probably not try to read this variable
directly.