BDFOY: Test fails with Business::ISBN installed [RT#33220]
Here's a better patch. The problem is that Business::ISBN 1.x used the unfortunate name
"country_code" which changed to the correct name "group_code" in Business::ISBN 2.x. Since
the new version of Business::ISBN handles ISBN-13, the EAN support is now just ISBN-13.

The trick is to support both major versions of the modules and Perl versions back to 5.004
(declared in Makefile.PL). I've tested this with Business::ISBN 1.84 and 2.03.