Treat ? as a reserved character in file: URIs
This seems to be the behaviour compatible with more current browsers.
Some additional rationale is found in this request:

---
Subject: URI: Query string do not work in file: URLs
From: Father Chrysostomos <sprout@cpan.org>
Message-Id: <0D380F9E-B2C1-4824-AE5E-A5B0A4A35E54@cpan.org>
Date: Sun, 15 Aug 2010 13:40:20 -0700
To: libwww@perl.org

URI::file has this little line in it:

sub path { shift->path_query(@_) }

which makes it impossible to test web pages offline (via file: rather
than http:). All web browsers I=92ve used allow a query string to be
added to a file URL, and require %3f to be used for files with question
marks in their names. The query string is not entirely useless. It
allows information to be passed to scripts embedded in the page.

Is there any chance this line could be removed?