Distingush between empty and undef values in query_form [RT#62708]
The _query module does not handle http query strings that are not
<name>=<value> pairs.

There is a behavior change. Previously, parameters sent into query_form
with value undef would be set to blank. Now they will be included with
no "=" or value. To set a value to blank, use an empty string -- ''.