URIs should be utf8::downgraded() internally [RT#70161]
This ensures that any parts extracted from the URL are returned as
non-UTF8 marked strings.  This might be less confusing to some
as you normally would need to decode("UTF-8", ...) parts you want to
treat as full Unicode strings.

Encode's decode_utf8() function fail to decode strings with the
UTF8 flag set, so this might also be seen as a workaround for that
issue.