Update the documentation for URI::_punycode to show better examples to fix GH Issue #45.
Standardized the documentation a bit as well by using head2 entries
instead of over and item. This allows metacpan and the like to
auto-generate a Table of Contents a bit more easily and makes direct
links to sections/functions a bit simpler.

The synopsis should be directly runnable now if someone were to
copy and paste it to test.